move directory file, fix cert var bug
