+++ packages/aserto/src/aserto/client/__init__.py
@@ -1,9 +1,11 @@
 from .options import AuthorizerOptions
 from .identity import Identity
 from .resource_context import ResourceContext
+from .directory import Directory
 
 __all__ = [
     "AuthorizerOptions",
+    "Directory",
     "Identity",
     "ResourceContext",
 ]
